+++ src/block_service/test/block_service_manager_test.cpp
@@ -33,7 +33,7 @@
 #include "block_service_mock.h"
 #include "metadata_types.h"
 #include "test_util/test_util.h"
-#include "utils/encryption_utils.h"
+#include "utils/env.h"
 #include "utils/error_code.h"
 #include "utils/filesystem.h"
 #include "utils/test_macros.h"
