+++ src/nfs/test/main.cpp
@@ -46,7 +46,7 @@
 #include "runtime/tool_api.h"
 #include "test_util/test_util.h"
 #include "utils/autoref_ptr.h"
-#include "utils/encryption_utils.h"
+#include "utils/env.h"
 #include "utils/error_code.h"
 #include "utils/filesystem.h"
 #include "utils/flags.h"
