+++ src/replica/bulk_load/replica_bulk_loader.cpp
@@ -46,7 +46,7 @@
 #include "utils/autoref_ptr.h"
 #include "utils/blob.h"
 #include "utils/chrono_literals.h"
-#include "utils/encryption_utils.h"
+#include "utils/env.h"
 #include "utils/fail_point.h"
 #include "utils/filesystem.h"
 #include "utils/fmt_logging.h"
