+++ src/replica/storage/simple_kv/simple_kv.server.impl.cpp
@@ -58,7 +58,7 @@
 #include "utils/autoref_ptr.h"
 #include "utils/binary_reader.h"
 #include "utils/blob.h"
-#include "utils/encryption_utils.h"
+#include "utils/env.h"
 #include "utils/filesystem.h"
 #include "utils/fmt_logging.h"
 #include "utils/utils.h"
