+++ src/server/pegasus_server_impl_init.cpp
@@ -51,7 +51,7 @@
 #include "server/pegasus_read_service.h"
 #include "server/pegasus_server_write.h" // IWYU pragma: keep
 #include "server/range_read_limiter.h"
-#include "utils/encryption_utils.h"
+#include "utils/env.h"
 #include "utils/flags.h"
 #include "utils/fmt_logging.h"
 #include "utils/strings.h"
