+++ src/utils/test/file_utils.cpp
@@ -33,7 +33,7 @@
 #include <string>
 #include <vector>
 
-#include "utils/encryption_utils.h"
+#include "utils/env.h"
 #include "utils/error_code.h"
 #include "utils/filesystem.h"
